requirements.txt
streamlit
langchain
langchain-text-splitters
langchain-google-genai
langchain-community
faiss-cpu
pypdf2
python-dotenv
pyyaml
bcrypt
